fetch/metadata/generated/form-submission-sec-fetch-user.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/cases/master.yml
- fetch/metadata/tools/templates/form-submission.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-User</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(method, url, userActivated) {
    const windowName = String(Math.random());
    const form = document.createElement('form');
    const submit = document.createElement('input');
    submit.setAttribute('type', 'submit');
    form.appendChild(submit);
    const win = open('about:blank', windowName);
    form.setAttribute('method', method);
    form.setAttribute('action', url);
    form.setAttribute('target', windowName);
    document.body.appendChild(form);

    // Query parameters must be expressed as form values so that they are sent
    // with the submission of forms whose method is POST.
    Array.from(new URL(url, location.origin).searchParams)
      .forEach(([name, value]) => {
        const input = document.createElement('input');
        input.setAttribute('type', 'hidden');
        input.setAttribute('name', name);
        input.setAttribute('value', value);
        form.appendChild(input);
      });

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });

        if (userActivated) {
          test_driver.click(submit);
        } else {
          submit.click();
        }
      })
      .then(() => {
        form.remove();
        win.close();
      });
  }
  const responseParams = {
    mime: 'text/html',
    body: `<script>opener.postMessage('done', '*')</${''}script>`
  };

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);
    const userActivated = false;
    return induceRequest('GET', url, userActivated)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'GET');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);
    const userActivated = true;
    return induceRequest('GET', url, userActivated)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-user');
          assert_equals(headers['sec-fetch-user'], '?1');
        });
  }, 'GET with user activation');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);
    const userActivated = false;
    return induceRequest('POST', url, userActivated)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'POST');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);
    const userActivated = true;
    return induceRequest('POST', url, userActivated)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-user');
          assert_equals(headers['sec-fetch-user'], '?1');
        });
  }, 'POST with user activation');
  </script>
  </body>
</html>
